sbc/requirements.txt
numpy==1.26.4
picamera2
opencv-python==4.9.0.80
tflite-runtime==2.14.0
pyserial
spidev
piexif
Pillow
sphinx
